changed variable types for blazor remove requirement

+++ C#/SEP3/FileData/DAOs/ProductFileDao.cs
@@ -16,7 +16,7 @@ namespace FileData.DAOs;
         
         
         
-        public Task DeleteAsync(int id)
+        public Task DeleteAsync(long id)
         {
             Product? existing = context.Products.FirstOrDefault(product => product.id == id);
             if (existing == null)
